requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2